Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -15,11 +15,11 @@ int main(void)
 	float d;
 	char c;
 
-	printf("Size of a char: %d byte(s)\n", sizeof(c));
-	printf("Size of an int: %d byte(s)\n", sizeof(i));
-	printf("Size of a long int: %ld byte(s)\n", sizeof(j));
-	printf("Size of a long long int: %lld byte(s)\n", sizeof(i));
-	printf("Size of a float: %f byte(s)\n", sizeof(i));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(j));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(i));
 	return (0);
 }
